requirements.txt
Flask==2.3.3
groq==0.4.1
PyPDF2==3.0.1
python-dotenv==1.0.0
tenacity==8.2.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3
gunicorn